+++ aspnet-core/src/OrderingSystem.Application/Customers/Dto/CreateCustomerDto.cs
@@ -7,7 +7,7 @@ namespace OrderingSystem.Customers.Dto
 {
     [AutoMapFrom(typeof(CustomerDto))]
     [AutoMapTo(typeof(Customer))]
-    public class CreateCustomerDto: EntityDto<int>
+    public class CreateCustomerDto: EntityDto<long>
     {
         public string Name { get; set; }
         public int DivisionId { get; set; } //FK
