+++ aspnet-core/src/OrderingSystem.Core/Entities/Customer.cs
@@ -3,7 +3,7 @@ using OrderingSystem.Authorization.Users;
 
 namespace OrderingSystem.Entities
 {
-    public class Customer: FullAuditedEntity<int>
+    public class Customer: FullAuditedEntity<long>
     {
         public string Name { get; set; }
         public int? DivisionId { get; set; }
